+++ src/DependencyInjection/Configuration.php
@@ -59,8 +59,8 @@ final class Configuration implements ConfigurationInterface
                     ->fixXmlConfig('ignore_transaction')
                     ->fixXmlConfig('prefix', 'prefixes')
                     ->children()
-                        ->arrayNode('integrations')
-                            ->scalarPrototype()->end()
+                        ->variableNode('integrations')
+                            ->defaultValue([])
                         ->end()
                         ->booleanNode('default_integrations')->end()
                         ->arrayNode('prefixes')
@@ -183,6 +183,7 @@ final class Configuration implements ConfigurationInterface
                     ->{interface_exists(MessageBusInterface::class) ? 'canBeDisabled' : 'canBeEnabled'}()
                     ->children()
                         ->booleanNode('capture_soft_fails')->defaultTrue()->end()
+                        ->booleanNode('isolate_breadcrumbs_by_message')->defaultFalse()->end()
                     ->end()
                 ->end()
             ->end();