+++ src/Tracing/Cache/TraceableCacheAdapterForV3WithNamespace.php
@@ -38,16 +38,12 @@ final class TraceableCacheAdapterForV3WithNamespace implements AdapterInterface,
      * {@inheritdoc}
      *
      * @param mixed[] $metadata
+     *
+     * @return mixed
      */
     public function get(string $key, callable $callback, ?float $beta = null, ?array &$metadata = null): mixed
     {
-        return $this->traceFunction('cache.get_item', function () use ($key, $callback, $beta, &$metadata) {
-            if (!$this->decoratedAdapter instanceof CacheInterface) {
-                throw new \BadMethodCallException(\sprintf('The %s::get() method is not supported because the decorated adapter does not implement the "%s" interface.', self::class, CacheInterface::class));
-            }
-
-            return $this->decoratedAdapter->get($key, $callback, $beta, $metadata);
-        }, $key);
+        return $this->traceGet($key, $callback, $beta, $metadata);
     }
 
     public function withSubNamespace(string $namespace): static