+++ CHANGELOG.md
@@ -1,8 +1,69 @@
 # CHANGELOG
 
+## 5.5.0
+
+The Sentry SDK team is happy to announce the immediate availability of Sentry Symfony SDK v5.5.0.
+
+### Features
+
+- Enable auto-instrumentation for the Symfony Cache Component [(#942)](https://github.com/getsentry/sentry-symfony/pull/942)
+- Add a new config flag to reset breadcrumbs between Symfony messages [(#946)](https://github.com/getsentry/sentry-symfony/pull/946)
+```yaml
+sentry:
+  messenger:
+    isolate_breadcrumbs_by_message: true
+```
+- Allow to pass in callbacks to configure SDK integrations [(#947)](https://github.com/getsentry/sentry-symfony/pull/947)
+```yaml
+services:
+    App\IntegrationCallback:
+        factory: ['App\IntegrationCallback', 'factory']
+
+sentry:
+    options:
+        integrations: 'App\IntegrationCallback'
+```
+
+## 5.4.1
+
+The Sentry SDK team is happy to announce the immediate availability of Sentry Symfony SDK v5.4.1.
+
+### Bug Fixes
+
+- Support named arguments in BufferFlushPass. [(#948)](https://github.com/getsentry/sentry-symfony/pull/948)
+
+## 5.4.0
+
+The Sentry SDK team is happy to announce the immediate availability of Sentry Symfony SDK v5.4.0.
+
+### Features
+
+- Introduce Sentry Structured Logging support. [(#940)](https://github.com/getsentry/sentry-symfony/pull/940)
+
+```yaml
+sentry:
+  options:
+    enable_logs: true
+
+services:
+  Sentry\SentryBundle\Monolog\LogsHandler:
+    arguments:
+      - !php/const Monolog\Logger::INFO
+
+monolog:
+  handlers:
+    sentry_logs:
+      type: service
+      id: Sentry\SentryBundle\Monolog\LogsHandler
+```
+
+### Bug Fixes
+
+- Add buffer flusher for sentry monolog handler. [(#936)](https://github.com/getsentry/sentry-symfony/pull/936)
+
 ## 5.3.1
 
-The Sentry SDK team is happy to announce the immediate availability of Sentry Symfony SDK v5.3.0.
+The Sentry SDK team is happy to announce the immediate availability of Sentry Symfony SDK v5.3.1.
 
 ### Bug Fixes